requirements.txt
faiss-cpu==1.8.0
hnswlib==0.8.0
numpy==1.24.4
pgvector==0.3.6
psycopg2==2.9.10
scikit-learn==1.3.2
scipy==1.10.1
tqdm==4.66.4
